+++ dashboard/src/app/(public)/publicUser/home/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+export default function PublicUserHomepage (){
+    return(
+        <div>
+            <h1>home page of public user</h1>
+        </div>
+    )
+
+}
