+++ dashboard/src/app/(public)/publicUser/marketplace/page.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function Marketplace (){
+    return(
+        <div>
+            <h1>Market plaace </h1>
+            <h1>Market plaace </h1>
+            <h1>Market plaace </h1>
+        </div>
+    )
+}
