--- dashboard/src/app/(public)/publicUsers/marketplace/page.tsx
@@ -1,15 +1,0 @@
-import PublicUsersNavigation from "@/app/Components/PublicUsers-Navigation";
-import React from "react";
-
-export default function Marketplace (){
-    return(
-        <div>
-            <header>
-                <PublicUsersNavigation />
-            </header>
-            <main>
-                <h1>MARKETPLACE</h1>
-            </main>
-        </div>
-    )
-}
