+++ dashboard/src/app/Components/Layout/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import Sidebar from "../Sidebar";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
- 
+
   return (
     <div className='flex min-h-screen'>
